% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electivity.R
\docType{package}
\name{electivity}
\alias{electivity}
\alias{electivity-package}
\title{Algorithms for electivity indices and measures of resource use versus availability.}
\description{
This package is basically Lechowicz 1982 turned into an R package, and includes
all algorithms that were described therein plus the example data that
was provided for gypsy moth resource utilisation.
}
\details{
Lechowicz, M.J., 1982. The sampling characteristics of electivity indices.
Oecologia 52, 22–30. https://doi.org/10.1007/BF00349007
}
\section{Author}{

Desi Quintans (@eco_desi)
}

\section{URL}{

\url{https://github.com/DesiQuintans/electivity/}
}

