% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funs.R
\name{ivlev_forage}
\alias{ivlev_forage}
\title{Ivlev's forage ratio, E'}
\usage{
ivlev_forage(r, p)
}
\arguments{
\item{r}{(Numeric) Resource utilisation.}

\item{p}{(Numeric) Resource availability.}
}
\value{
A numeric vector.
}
\description{
Ivlev's forage ratio, E'
}
\examples{
data(moth_distrib)
ivlev_forage(moth_distrib$r, moth_distrib$p)
}
