% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funs.R
\name{jacob_electivity}
\alias{jacob_electivity}
\title{Jacob's modified electivity, D}
\usage{
jacob_electivity(r, p)
}
\arguments{
\item{r}{(Numeric) Resource utilisation.}

\item{p}{(Numeric) Resource availability.}
}
\value{
A numeric vector.
}
\description{
Bounded between +0.1 (avoidance), +1.0 (random feeding), and infinity (preference).
}
\examples{
data(moth_distrib)
jacob_electivity(moth_distrib$r, moth_distrib$p)
}
