% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funs.R
\name{strauss_linear}
\alias{strauss_linear}
\title{Strauss' linear index, L}
\usage{
strauss_linear(r, p)
}
\arguments{
\item{r}{(Numeric) Resource utilisation.}

\item{p}{(Numeric) Resource availability.}
}
\value{
A numeric vector.
}
\description{
Strauss' linear index, L
}
\examples{
data(moth_distrib)
strauss_linear(moth_distrib$r, moth_distrib$p)
}
