% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funs.R
\name{vs_electivity}
\alias{vs_electivity}
\title{Vanderploeg and Scavia's relativised electivity, E*}
\usage{
vs_electivity(r, p, na.rm = TRUE)
}
\arguments{
\item{r}{(Numeric) Resource utilisation.}

\item{p}{(Numeric) Resource availability.}

\item{na.rm}{(Logical) If \code{TRUE}, \code{NA}s will be ignored when calculating the
selectivity coefficient (W).}
}
\value{
A numeric vector.
}
\description{
Vanderploeg and Scavia's relativised electivity, E*
}
\examples{
data(moth_distrib)
vs_electivity(moth_distrib$r, moth_distrib$p)

}
