% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funs.R
\name{chesson_alpha}
\alias{chesson_alpha}
\alias{vs_select_coef}
\title{Chesson's alpha, or Vanderploeg and Scavia's selectivity coefficient (W)}
\usage{
chesson_alpha(r, p)

vs_select_coef(r, p)
}
\arguments{
\item{r}{(Numeric) Resource utilisation.}

\item{p}{(Numeric) Resource availability.}
}
\value{
A numeric vector.
}
\description{
Chesson's alpha, or Vanderploeg and Scavia's selectivity coefficient (W)
}
\examples{
data(moth_distrib)
chesson_alpha(moth_distrib$r, moth_distrib$p)

\dontrun{
vs_select_coef(moth_distrib$r, moth_distrib$p)
}
}
