% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funs.R
\name{ivlev_electivity}
\alias{ivlev_electivity}
\title{Ivlev's electivity, E}
\usage{
ivlev_electivity(r, p)
}
\arguments{
\item{r}{(Numeric) Resource utilisation.}

\item{p}{(Numeric) Resource availability.}
}
\value{
A numeric vector.
}
\description{
Ivlev's electivity, E
}
\examples{
data(moth_distrib)
ivlev_electivity(moth_distrib$r, moth_distrib$p)
}
