% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funs.R
\name{jacob_forage}
\alias{jacob_forage}
\title{Jacob's modified forage ratio, Q (log10)}
\usage{
jacob_forage(r, p, log = TRUE)
}
\arguments{
\item{r}{(Numeric) Resource utilisation.}

\item{p}{(Numeric) Resource availability.}

\item{log}{(Logical) If TRUE, return the value as Log10.}
}
\value{
A numeric vector.
}
\description{
Jacob's modified forage ratio, Q (log10)
}
\examples{
data(moth_distrib)
jacob_forage(moth_distrib$r, moth_distrib$p, log = TRUE)
jacob_forage(moth_distrib$r, moth_distrib$p, log = FALSE)
}
