% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{moth_distrib}
\alias{moth_distrib}
\title{Distribution of gypsy moth larvae, Lymantria dispar, feeding in a deciduous forest
in southwestern Quebec, Canada.}
\format{An object of class \code{data.frame} with 19 rows and 6 columns.}
\usage{
moth_distrib
}
\description{
This is Table 2 in the cited source paper.
}
\section{Author}{

Martin Lechowicz
}

\section{Source}{

Lechowicz, M.J., 1982. The sampling characteristics of electivity indices.
Oecologia 52, 22–30. https://doi.org/10.1007/BF00349007
}

\examples{
# data(moth_distrib)

}
\keyword{data}
